+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
-## [0.0.1]
+All notable changes to the "amstrad-basic-helper-commands" extension will be documented in this file.
 
-- Initial release.
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+
+## [Unreleased]
+
+- Initial release